app/requirements.txt
Flask
mysql-connector-python
requests
BeautifulSoup4
lxml